Makefile: modify compilation of schemas
By default, executing the Makefile of schemas directory generates the
binpb files only. Since we need the C files, modify the Makefile to
generate c and h files from schemas/*.proto files using protoc directly
in the main Makefile. Add these files to the compilation process by
modifying the OBJ variable and adding the required flags and libraries
using pkg-config.

The generated C and H files are placed in build/schemas directory as
they are the result of a compilation and are not intended to be
modified by the user.

SFL: #21154
Change-Id: I71574514f1e587b3d89aa79c5f6aa5d5e8f84419
Signed-off-by: Kévin L'hôpital <kevin.lhopital@savoirfairelinux.com>