gapi_mosquitto: don't exit program after sending message or on error
As the API program is a daemon, it should not exit after sending
a message or after and error. Instead, it should just continue running.
Then remove the exit(1) call in on_publish callback when the mid does
not match the expected one and remove the set of running to false that
will stop the main loop after sending a message.

SFL: #21154
Change-Id: I2f7bfe1ac2a8a5cd1830013cecfd59826ae1609c
Signed-off-by: Kévin L'hôpital <kevin.lhopital@savoirfairelinux.com>